% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSentence.R
\name{createSentence}
\alias{createSentence}
\title{Function to create a sentence from individual words}
\usage{
createSentence(subject, adverb, verb, noun)
}
\arguments{
\item{subject}{The subject of the sentence. Choices are limited to 'I',
'You', 'She', 'He', 'They', and 'We'.}

\item{adverb}{An optional word to modify the verb of the sentence.}

\item{verb}{The word indicating what action is happening in the sentence.}

\item{noun}{The word indicating the object to which the action is happening.}
}
\value{
Returns a complete sentence as a single string.
}
\description{
Setting up an R package is easy when using the RStudio GUI as
as well as the R packages \code{devtools}, \code{usethis}, \code{Rdpack},
\code{testhat}, and \code{roxygen}. You can nicely document your functions,
and easily share argument definitions across multiple functions with the
roxygen \code{@template} tag. We can also make our documentation fancy by
adding math, for example, like this: \eqn{\sum_{i}^{n} x^i}.
}
\examples{
subject = "I"
adverb = "happily"
verb = "walked"
noun = "home"
createSentence(subject, adverb, verb, noun)
}
\references{
\insertRef{wickham2015r}{learnerpkg}
}
